fix/atualizando-estrutura: atualizando a estrutura do projeto que estava com o nome errado
