+++ app/src/main/java/com/example/igestamobile/Perfil.java
@@ -1,4 +1,4 @@
-package com.matheus.igestamobile;
+package com.example.igestamobile;
 
 import android.os.Bundle;
 
@@ -7,13 +7,14 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.LinearLayout;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link Configuracao#newInstance} factory method to
+ * Use the {@link Perfil#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Configuracao extends Fragment {
+public class Perfil extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +25,7 @@ public class Configuracao extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public Configuracao() {
+    public Perfil() {
         // Required empty public constructor
     }
 
@@ -34,11 +35,11 @@ public class Configuracao extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Configuracao.
+     * @return A new instance of fragment Perfil.
      */
     // TODO: Rename and change types and number of parameters
-    public static Configuracao newInstance(String param1, String param2) {
-        Configuracao fragment = new Configuracao();
+    public static Perfil newInstance(String param1, String param2) {
+        Perfil fragment = new Perfil();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +60,22 @@ public class Configuracao extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_configuracao, container, false);
+
+        View view = inflater.inflate(R.layout.fragment_perfil, container, false);
+
+        LinearLayout bt_config = view.findViewById(R.id.bt_config);
+
+        bt_config.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Configuracao configuracao = new Configuracao();
+                getActivity().getSupportFragmentManager()
+                        .beginTransaction()
+                        .replace(R.id.container, configuracao)
+                        .commit();
+            }
+        });
+
+        return view;
     }
 }
