+++ app/src/main/java/com/example/igestamobile/ui/notifications/NotificationsFragment.java
@@ -1,4 +1,4 @@
-package com.matheus.igestamobile.ui.notifications;
+package com.example.igestamobile.ui.notifications;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,7 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.matheus.igestamobile.databinding.FragmentNotificationsBinding;
+import com.example.igestamobile.databinding.FragmentNotificationsBinding;
 
 public class NotificationsFragment extends Fragment {
