+++ app/src/test/java/com/example/igestamobile/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.matheus.igestamobile;
+package com.example.igestamobile;
 
 import org.junit.Test;
